SVM: add `program_loader` module
